requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
arrow==1.4.0
asgiref==3.11.0
blessed==1.25.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
cloudscraper==1.2.71
cryptography==45.0.6
Django==5.2.8
django-cors-headers==4.9.0
django-allauth==65.13.0
django-appconf==1.2.0
django-debug-toolbar==6.0.0
django-encrypted-model-fields==0.6.5
django-picklefield==3.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jinxed==1.3.0
psycopg2-binary==2.9.10
oauthlib==3.3.1
pillow==11.3.0
py-vapid==1.9.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==3.5.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
